Add fun stats
This was my first time working with generics. I have no idea whether or not it's good, but it works. (at least, for longs. I have no idea what will happen the moment I add another type that can be added. if it breaks i'll fix it when the time comes)